refactor(buildings): hoist pack loop and composition into placer
Move `pack_kinds` / `PackHost` and corner-L / peninsula helpers under
`crate::placer`, add host-relative `ProposeKnobs` + `SoftGoalRole::Appointed`,
share `FurnitureFill`, and thin kitchen / CommonBedroom onto those APIs.

## Agent Dialogue

- **Developer:** Do the worth-fixing items and smaller cleanups from the retro; then refactor kitchen a bit.
- **Agent:** Hoisted pack + composition into placer; FreeExtentFrac / WallLongFrac; SoftGoalRole-driven soft-goal; shared FurnitureFill; kitchen uses try_corner_l / peninsula; bedroom storage via propose_from_spec; gallery smoke → `livable_quarters/tests.rs`. `usage_areas::` tests green.
