+++ maybraid/richmond/buildings/src/usage_areas/livable_quarters/study/layout.rs
@@ -5,10 +5,9 @@ use procedural_common::NoiseParams;
 
 use crate::fit::{Confines, FitError};
 use crate::placer::{
-	CommitEffect, KindSpec, Predicate, ProgramTier, ProposeKnobs, SoftGoalRole,
+	init_host, pack_kinds, CommitEffect, KindSpec, PackKnobs, Predicate, ProgramTier, ProposeKnobs,
+	SoftGoalRole,
 };
-
-use crate::usage_areas::livable_quarters::pack::{init_host, pack_kinds, PackKnobs};
 
 pub const MIN_AREA: f32 = 2.0 * 2.0;
 
@@ -38,10 +37,10 @@ impl StudyRegions {
 				tier: ProgramTier::Appointed,
 				weight: 1.0,
 				max_count: Some(1),
-				soft_goal: SoftGoalRole::None,
-				propose: ProposeKnobs::WallLong {
-					along_min: 1.0,
-					along_max: 1.8,
+				soft_goal: SoftGoalRole::Appointed,
+				propose: ProposeKnobs::WallLongFrac {
+					along_frac_min: 0.25,
+					along_frac_max: 0.45,
 					depth_min: 0.55,
 					depth_max: 0.75,
 					height: 0.75,
@@ -56,9 +55,9 @@ impl StudyRegions {
 				weight: 0.55,
 				max_count: Some(2),
 				soft_goal: SoftGoalRole::None,
-				propose: ProposeKnobs::WallLong {
-					along_min: 0.7,
-					along_max: 1.2,
+				propose: ProposeKnobs::WallLongFrac {
+					along_frac_min: 0.14,
+					along_frac_max: 0.24,
 					depth_min: 0.35,
 					depth_max: 0.5,
 					height: 1.8,
@@ -90,7 +89,6 @@ impl StudyRegions {
 			&mut host,
 			confines,
 			noise,
-			|p| p.iter().any(|(k, _)| *k == StudyKind::Desk),
 		)?;
 		for (kind, aabb) in placed {
 			match kind {
